+++ lib/src/components/components.dart
@@ -23,6 +23,7 @@ export 'segmented_control/octo_segmented_control.dart';
 export 'segmented_control/octo_segmented_control_item.dart';
 export 'skeleton/octo_skeleton.dart';
 export 'spinner/octo_spinner.dart';
+export 'state_label/octo_state_label.dart';
 export 'switch/octo_switch.dart';
 export 'text_field/octo_text_field.dart';
 export 'toast/octo_toast.dart';